pyproject.toml
[tool.poetry]
name = "sentiment-api"
version = "0.1.0"
description = "API de Análise de Sentimentos para textos"
authors = ["Thiago Di Faria <sthiagodifaria@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.20"
pydantic = "^2.3.0"
python-dotenv = "^1.0.0"
nltk = "^3.8.1"
textblob = "^0.17.1"
pandas = "^2.1.0"
spacy = "^3.7.2"
fasttext-wheel = "^0.9.2"
langdetect = "^1.0.9"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
black = "^23.7.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88